Allow systemd to provide capabilities
With this change, we have xrootd start as root and then drop privileges
to the xrootd user.  This allows the service to inherit capabilities
provided by systemd on RHEL7 (since RHEL7 systemd doesn't support ambient
capabilities).

This commit also includes code to test for permitted, but not effective,
capability settings (default for systemd) and set the effective capabilities
accordingly.

New systemd unit provided in addition to updated README.