Make multiuser plugin compatible with Macaroons.
The multiuser plugin would use AOP_Any with path / to force the underlying
authorization object to generate a username for the session.  However,
this doesn't work for token-based authorization as the name might only
be known along with a particular path.  Further, the Macaroons plugin
refuses to map with AOP_Any to prevent a macaroon from being used to
generate a macaroon.